requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pylint==2.17.0
ibm-watson==6.1.0